OHRM5X-1856: Refactor code
