+++ installer/Command/InstallOnExistingDatabaseCommand.php
@@ -17,7 +17,7 @@
  * Boston, MA 02110-1301, USA
  */
 
-namespace OrangeHRM\Installer\cli;
+namespace OrangeHRM\Installer\Command;
 
 use OrangeHRM\Authentication\Dto\UserCredential;
 use OrangeHRM\Installer\Util\AppSetupUtility;
